requirements.txt
click==7.0
gunicorn==19.9.0
flask==1.0.2
itsdangerous==1.1.0
jinja2==2.10.3
markupsafe==1.1.1
pillow==6.2.0
pytesseract==0.2.6
werkzeug==0.16.0
cv2==4.1.1
re==2.2.1
